refactor: move notice files into feature-based structure
