+++ lib/domain/notice/entites/notice_entity.dart
